Add dropped chan
This chan signals that conn decided to drop the connection.
A Torrent should receive from this chan.We also send a
connDropped message when droping conns
but there is a problem if only the latter one  exists.
We  can't ensure that commandCh doesn't deadlock
when we drop conn (maybe we'll send many commands
from Torrent but conn won't receive them).