Add new pieces impl
The previous one was overcomplicated. Now when a conn
wants to send requests, it just acquires a lock and takes them.
When it want  to discard requests,it acquires the lock , drops
them and the remaining conns are notified that requests are
available.Moreover, every piece has a rarity field which gets
updated when have or bitfield messages are received by conns.
We still prioritize partially requested pieces regardless of our
request strategy.