Refactor dial,listen and accept & Add xported methods
Now we use a more idiomatic way to dial and accept bt
connections by implementing dialer and listener.

Also, if a connection we can't inform a Torrent about
its presence before 5 seconds, then we close the connection.

In addition, handshake logic to check if we are interested
in the infohash the peer send us was moved in the torrent pkg.

Lastly, I introduced a parser interface for the MetaInfo and
some implementations of it (useful for abstracting the input
source).